Add functions to spiffeutils
